➡️ Refactored the result sorting method Moved it into its own class.
